+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   Replace custom delay method 'Wait' with 'C_Timer.After'
+
 ## [1.1.0] - 2024-09-05
 
 ### Added